Move logic for auth from index to a new file
