+++ web/src/firebase-components/auth.js
@@ -1,0 +1,139 @@
+// Signs-in Friendly Chat.
+import {getAuth, GoogleAuthProvider, onAuthStateChanged, signInWithPopup, signOut} from "firebase/auth";
+import {getMessaging, getToken, onMessage} from "firebase/messaging";
+import {doc, getFirestore, setDoc} from "firebase/firestore";
+
+export async function signIn() {
+    // Sign in Firebase using popup auth and Google as the identity provider.
+    var provider = new GoogleAuthProvider();
+    await signInWithPopup(getAuth(), provider);
+}
+
+// Signs-out of Friendly Chat.
+export function signOutUser() {
+    // Sign out of Firebase.
+    // console.log('out');
+    signOut(getAuth());
+}
+
+// Initialize firebase auth
+export function initFirebaseAuth() {
+    // Listen to auth state changes.
+    onAuthStateChanged(getAuth(), authStateObserver);
+}
+
+// Returns the signed-in user's profile Pic URL.
+export function getProfilePicUrl() {
+    return getAuth().currentUser.photoURL || '/images/profile_placeholder.png';
+}
+
+// Returns the signed-in user's display name.
+export function getUserName() {
+    return getAuth().currentUser.displayName;
+}
+
+// Returns true if a user is signed-in.
+export function isUserSignedIn() {
+    return !!getAuth().currentUser;
+} // Adds a size to Google Profile pics URLs.
+// Saves the messaging device token to Cloud Firestore.
+async function saveMessagingDeviceToken() {
+    try {
+        const currentToken = await getToken(getMessaging());
+        if (currentToken) {
+            console.log('Got FCM device token:', currentToken);
+            // Saving the Device Token to Cloud Firestore.
+            const tokenRef = doc(getFirestore(), 'fcmTokens', currentToken);
+            await setDoc(tokenRef, {uid: getAuth().currentUser.uid});
+
+            // This will fire when a message is received while the app is in the foreground.
+            // When the app is in the background, firebase-messaging-sw.js will receive the message instead.
+            onMessage(getMessaging(), (message) => {
+                console.log(
+                    'New foreground notification from Firebase Messaging!',
+                    message.notification
+                );
+            });
+        } else {
+            // Need to request permissions to show notifications.
+            requestNotificationsPermissions();
+        }
+    } catch (error) {
+        console.error('Unable to get messaging token.', error);
+    }
+    ;
+}
+
+// Requests permissions to show notifications.
+async function requestNotificationsPermissions() {
+    console.log('Requesting notifications permission...');
+    const permission = await Notification.requestPermission();
+
+    if (permission === 'granted') {
+        console.log('Notification permission granted.');
+        // Notification permission granted.
+        await saveMessagingDeviceToken();
+    } else {
+        console.log('Unable to get permission to notify.');
+    }
+}
+
+// Triggers when the auth state change for instance when the user signs-in or signs-out.
+function authStateObserver(user) {
+    if (user) { // User is signed in!
+        // Get the signed-in user's profile pic and name.
+        var profilePicUrl = getProfilePicUrl();
+        var userName = getUserName();
+
+        // Set the user's profile pic and name.
+        userPicElement.style.backgroundImage = 'url(' + addSizeToGoogleProfilePic(profilePicUrl) + ')';
+        userNameElement.textContent = userName;
+
+        // Show user's profile and sign-out button.
+        userNameElement.removeAttribute('hidden');
+        userPicElement.removeAttribute('hidden');
+        signOutButtonElement.removeAttribute('hidden');
+
+        // Hide sign-in button.
+        signInButtonElement.setAttribute('hidden', 'true');
+
+        // We save the Firebase Messaging Device token and enable notifications.
+        saveMessagingDeviceToken();
+    } else { // User is signed out!
+        // Hide user's profile and sign-out button.
+        userNameElement.setAttribute('hidden', 'true');
+        userPicElement.setAttribute('hidden', 'true');
+        signOutButtonElement.setAttribute('hidden', 'true');
+
+        // Show sign-in button.
+        signInButtonElement.removeAttribute('hidden');
+    }
+}
+
+export function addSizeToGoogleProfilePic(url) {
+    if (url.indexOf('googleusercontent.com') !== -1 && url.indexOf('?') === -1) {
+        return url + '?sz=150';
+    }
+    return url;
+}
+
+var userPicElement = document.getElementById('user-pic');
+var userNameElement = document.getElementById('user-name');
+export var signInButtonElement = document.getElementById('sign-in');
+export var signOutButtonElement = document.getElementById('sign-out'); // Returns true if user is signed-in. Otherwise false and displays a message.
+export function checkSignedInWithMessage() {
+    // Return true if the user is signed in Firebase
+    if (isUserSignedIn()) {
+        return true;
+    }
+
+    // Display a message to the user using a Toast.
+    var data = {
+        message: 'You must sign-in first',
+        timeout: 2000
+    };
+    signInSnackbarElement.MaterialSnackbar.showSnackbar(data);
+    return false;
+}
+
+export var signInSnackbarElement = document.getElementById('must-signin-snackbar');
