+++ web/src/index.js
@@ -15,77 +15,37 @@
  */
  'use strict';
 
- import { initializeApp } from 'firebase/app';
- import {
-   getAuth,
-   onAuthStateChanged,
-   GoogleAuthProvider,
-   signInWithPopup,
-   signOut,
- } from 'firebase/auth';
- import {
-   getFirestore,
-   collection,
-   addDoc,
-   query,
-   orderBy,
-   limit,
-   onSnapshot,
-   setDoc,
-   updateDoc,
-   doc,
-   serverTimestamp,
- } from 'firebase/firestore';
- import {
-   getStorage,
-   ref,
-   uploadBytesResumable,
-   getDownloadURL,
- } from 'firebase/storage';
- import {
-   getMessaging,
-   getToken,
-   onMessage
- } from 'firebase/messaging'; 
- import { getPerformance } from 'firebase/performance';
+import {initializeApp} from 'firebase/app';
+import {getAuth,} from 'firebase/auth';
+import {
+    addDoc,
+    collection,
+    getFirestore,
+    limit,
+    onSnapshot,
+    orderBy,
+    query,
+    serverTimestamp,
+    updateDoc,
+} from 'firebase/firestore';
+import {getDownloadURL, getStorage, ref, uploadBytesResumable,} from 'firebase/storage';
+import {getPerformance} from 'firebase/performance';
 
- import { getFirebaseConfig } from './firebase-config.js';
- 
- // Signs-in Friendly Chat.
- async function signIn() {
-   // Sign in Firebase using popup auth and Google as the identity provider.
-   var provider = new GoogleAuthProvider();
-   await signInWithPopup(getAuth(), provider);
- }
- 
- // Signs-out of Friendly Chat.
- function signOutUser() {
-   // Sign out of Firebase.
-   signOut(getAuth());
- }
- 
- // Initialize firebase auth
- function initFirebaseAuth() {
-   // Listen to auth state changes.
-   onAuthStateChanged(getAuth(), authStateObserver);
- }
- 
- // Returns the signed-in user's profile Pic URL.
- function getProfilePicUrl() {
-   return getAuth().currentUser.photoURL || '/images/profile_placeholder.png';
- }
- 
- // Returns the signed-in user's display name.
- function getUserName() {
-   return getAuth().currentUser.displayName;
- }
- 
- // Returns true if a user is signed-in.
- function isUserSignedIn() {
-   return !!getAuth().currentUser;
- }
- 
- // Saves a new message to Cloud Firestore.
+import {getFirebaseConfig} from './firebase-config.js';
+import {
+    addSizeToGoogleProfilePic,
+    checkSignedInWithMessage,
+    getProfilePicUrl,
+    getUserName,
+    initFirebaseAuth,
+    signIn,
+    signInButtonElement,
+    signInSnackbarElement,
+    signOutButtonElement,
+    signOutUser
+} from "./firebase-components/auth";
+
+// Saves a new message to Cloud Firestore.
  async function saveMessage(messageText) {
    // Add a new message entry to the Firebase database.
    try {
@@ -150,47 +110,6 @@
    }
  }
  
- // Saves the messaging device token to Cloud Firestore.
- async function saveMessagingDeviceToken() {
-   try {
-     const currentToken = await getToken(getMessaging());
-     if (currentToken) {
-       console.log('Got FCM device token:', currentToken);
-       // Saving the Device Token to Cloud Firestore.
-       const tokenRef = doc(getFirestore(), 'fcmTokens', currentToken);
-       await setDoc(tokenRef, { uid: getAuth().currentUser.uid });
-
-       // This will fire when a message is received while the app is in the foreground.
-       // When the app is in the background, firebase-messaging-sw.js will receive the message instead.
-       onMessage(getMessaging(), (message) => {
-         console.log(
-           'New foreground notification from Firebase Messaging!',
-           message.notification
-         );
-       });
-     } else {
-       // Need to request permissions to show notifications.
-       requestNotificationsPermissions();
-     }
-   } catch(error) {
-     console.error('Unable to get messaging token.', error);
-   };
- }
- 
- // Requests permissions to show notifications.
- async function requestNotificationsPermissions() {
-   console.log('Requesting notifications permission...');
-   const permission = await Notification.requestPermission();
-   
-   if (permission === 'granted') {
-     console.log('Notification permission granted.');
-     // Notification permission granted.
-     await saveMessagingDeviceToken();
-   } else {
-     console.log('Unable to get permission to notify.');
-   }
- }
- 
  // Triggered when a file is selected via the media picker.
  function onMediaFileSelected(event) {
    event.preventDefault();
@@ -227,54 +146,6 @@
    }
  }
  
- // Triggers when the auth state change for instance when the user signs-in or signs-out.
- function authStateObserver(user) {
-   if (user) { // User is signed in!
-     // Get the signed-in user's profile pic and name.
-     var profilePicUrl = getProfilePicUrl();
-     var userName = getUserName();
- 
-     // Set the user's profile pic and name.
-     userPicElement.style.backgroundImage = 'url(' + addSizeToGoogleProfilePic(profilePicUrl) + ')';
-     userNameElement.textContent = userName;
- 
-     // Show user's profile and sign-out button.
-     userNameElement.removeAttribute('hidden');
-     userPicElement.removeAttribute('hidden');
-     signOutButtonElement.removeAttribute('hidden');
- 
-     // Hide sign-in button.
-     signInButtonElement.setAttribute('hidden', 'true');
- 
-     // We save the Firebase Messaging Device token and enable notifications.
-     saveMessagingDeviceToken();
-   } else { // User is signed out!
-     // Hide user's profile and sign-out button.
-     userNameElement.setAttribute('hidden', 'true');
-     userPicElement.setAttribute('hidden', 'true');
-     signOutButtonElement.setAttribute('hidden', 'true');
- 
-     // Show sign-in button.
-     signInButtonElement.removeAttribute('hidden');
-   }
- }
- 
- // Returns true if user is signed-in. Otherwise false and displays a message.
- function checkSignedInWithMessage() {
-   // Return true if the user is signed in Firebase
-   if (isUserSignedIn()) {
-     return true;
-   }
- 
-   // Display a message to the user using a Toast.
-   var data = {
-     message: 'You must sign-in first',
-     timeout: 2000
-   };
-   signInSnackbarElement.MaterialSnackbar.showSnackbar(data);
-   return false;
- }
- 
  // Resets the given MaterialTextField.
  function resetMaterialTextfield(element) {
    element.value = '';
@@ -288,14 +159,6 @@
        '<div class="message"></div>' +
        '<div class="name"></div>' +
      '</div>';
- 
- // Adds a size to Google Profile pics URLs.
- function addSizeToGoogleProfilePic(url) {
-   if (url.indexOf('googleusercontent.com') !== -1 && url.indexOf('?') === -1) {
-     return url + '?sz=150';
-   }
-   return url;
- }
  
  // A loading image URL.
  var LOADING_IMAGE_URL = 'https://www.google.com/images/spin-32.gif?a';
@@ -398,11 +261,6 @@
  var imageButtonElement = document.getElementById('submitImage');
  var imageFormElement = document.getElementById('image-form');
  var mediaCaptureElement = document.getElementById('mediaCapture');
- var userPicElement = document.getElementById('user-pic');
- var userNameElement = document.getElementById('user-name');
- var signInButtonElement = document.getElementById('sign-in');
- var signOutButtonElement = document.getElementById('sign-out');
- var signInSnackbarElement = document.getElementById('must-signin-snackbar');
  
  // Saves message on form submit.
  messageFormElement.addEventListener('submit', onMessageFormSubmit);
